Rename navigations tests models folder SN-822
