+++ src/Saritasa.NetForge.Tests/EfCoreMetadataServiceTests/NavigationMetadata/Models/Client.cs
@@ -1,11 +1,11 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Saritasa.NetForge.Tests.EfCoreMetadataServiceTests.NavigationMetadata.Entities;
+namespace Saritasa.NetForge.Tests.EfCoreMetadataServiceTests.NavigationMetadata.Models;
 
 /// <summary>
-/// Project.
+/// Client.
 /// </summary>
-public class Project
+public class Client
 {
     /// <summary>
     /// Identifier.
@@ -17,9 +17,4 @@ public class Project
     /// </summary>
     [MaxLength(100)]
     public string Name { get; set; } = null!;
-
-    /// <summary>
-    /// The client who is owner of the project.
-    /// </summary>
-    public Client Client { get; set; } = null!;
 }
